Improve documentation; Move RestrictedMapboxSearchAPI to public sdk-common module
